src/requirements.txt
sentry-sdk
python-dotenv
paramiko
loguru
boto3
